Move story about elements as field labels in Datagrid
